components/statistics: Переносит методы из filter.js. Правит ошибку целочисленного деления для продолжительности фильма
