+++ src/controllers/filter.js
@@ -1,9 +1,7 @@
 import FilterComponent, {MenuItem} from '../components/main-menu.js';
 import {FilterType} from '../const.js';
 import {render, replace, RenderPosition} from '../utils/render.js';
-import {getMoviesByFilter, getWatchedMovies} from '../utils/filter.js';
-import {GenreItems} from '../const.js';
-import {getFormattedRuntime} from '../utils/common.js';
+import {getMoviesByFilter} from '../utils/filter.js';
 
 export default class FilterController {
   constructor(container, moviesModel, statisticsComponent, pageController) {
@@ -13,7 +11,6 @@ export default class FilterController {
     this._pageController = pageController;
 
     this._movies = this._moviesModel.getMovies();
-    this._watchedMovies = getWatchedMovies(this._movies);
 
     this._activeFilterType = FilterType.ALL;
     this._filterComponent = null;
@@ -47,50 +44,6 @@ export default class FilterController {
     } else {
       render(container, this._filterComponent.getElement(), RenderPosition.BEFOREEND);
     }
-
-    console.log(`all watched movies`, this._watchedMovies);
-
-    console.table(this._getStatisticsData());
-
-    console.table(this._getTotalMoviesRuntime());
-
-  }
-
-  _getTotalMoviesRuntime() {
-    const watchedMovies = this._watchedMovies;
-    let totalMinutes = 0;
-    watchedMovies.forEach((movie) => {
-      totalMinutes += movie.duration;
-    });
-
-    const totalRuntime = getFormattedRuntime(totalMinutes);
-
-    return totalRuntime.digits;
-  }
-
-  _getStatisticsData() {
-    let moviesStatistics = [];
-    GenreItems.map((genre) => {
-      let count = this._getMoviesByGenre(genre).length;
-      moviesStatistics.push(this._getMovieStatisticElement(genre, count));
-    });
-
-    let mostWatchedGenreList = moviesStatistics.sort((a, b) => b.movieCount - a.movieCount);
-    mostWatchedGenreList = mostWatchedGenreList.slice(0, 5);
-
-    return mostWatchedGenreList;
-  }
-
-  _getMovieStatisticElement(genre, count) {
-    return {
-      label: genre,
-      movieCount: count,
-    };
-  }
-
-  _getMoviesByGenre(genre) {
-    const moviesByGenre = this._watchedMovies.filter((film) => film.genres.includes(genre));
-    return moviesByGenre;
   }
 
   _onDataChange() {
